modified signup.php
A check or validation was put in place to avoid duplicate email address, since our email address serve as our unique entity

+++ signup.php
@@ -10,16 +10,33 @@ if(isset($_POST['signupBtn'])){
 	$formArray = $_POST; #retrieves the form values passed
 	$formArray['password'] = password_hash($formArray['password'], PASSWORD_DEFAULT); #Hash the password
 	
-	$data['id'] = 1;
-	$data = array_merge($data, $formArray); #combine id to the data from form
-
-	#Removes the SignUpBtn since it is returning an empty value
-    unset($data['signupBtn']);
-
-	$json_data = json_encode($data); #Encode data to JSON/Object format
+	$json_retrieved_data = file_get_contents($fileDB);
+    	$decoded_data = json_decode($json_retrieved_data, true);
 	
-	/**Every new registration should be appended (i.e added to the next line of the previous)*/
-	file_put_contents($fileDB, $json_data, FILE_APPEND);
-}
+    	$email = $formArray['email'];
+    	$email_exists = false;
+		
+	foreach ($decoded_data as $key => $value) {
+		if ($value['email'] == $email) {
+		    $email_exists = true;
+		    break;
+		}
+	}
+	if ($email_exist) {
+		Location('index.html');
+	  } 
+	else {
+	
+		//$data['id'] = 1;
+		//$data = array_merge($data, $formArray); #combine id to the data from form
 
+		#Removes the SignUpBtn since it is returning an empty value
+		unset($formArray['signupBtn']);
+
+		$json_data = json_encode($formArray); #Encode data to JSON/Object format
+
+		/**Every new registration should be appended (i.e added to the next line of the previous)*/
+		file_put_contents($fileDB, $json_data, FILE_APPEND);
+	}		
+}
 ?>
